# Improve performance of memResize().